# Add native UTF-8 Validation using fast shift based DFA