--- a/base/strings/string.jl
+++ b/base/strings/string.jl
@@ -218,7 +218,7 @@ end
 
 _isvalid_utf8(s::Union{String,Vector{UInt8},FastContiguousSubArray{UInt8,1,Vector{UInt8}}}) = _isvalid_utf8(unsafe_wrap(Vector{UInt8}, s))
 
-#Classifcations of string
+# Classifcations of string
     # 0: neither valid ASCII nor UTF-8
     # 1: valid ASCII
     # 2: valid UTF-8