This is a based on the discussion in #41533.  It is a julia implementation of a shift based DFA implemented with inspiration from https://github.com/golang/go/issues/47120.


benchmarking reveals that the isvalid method present in 1.8.3 using the c code can achieve:
UTF-8: isvalid 1.4413834707420068 billion bytes/second
ASCII: isvalid 1.56482109432222 billion bytes/second

While the new julia version is even faster at:
UTF-8: isvalid 2.346150730742082 billion bytes/second
ASCII: isvalid 2.1939956860500947 billion bytes/second




<details>
  <summary>Code used for benchmarking</summary>

```julia
using BenchmarkTools
using Statistics
const utf_base_string = "0123456789日本語日本語日本語日abcdefghijklmnopqrstuvwx"
const ascii_base_string = "0123456789abcdefghijklmnopqrstuvwx0123456789"

function benchmark_utfvalid(fun; string_length_bytes = 1024)
    repeats = div(string_length_bytes, sizeof(utf_base_string))
    test_string = repeat(utf_base_string,repeats)
    num = sizeof(test_string)
    t = @benchmark state=$fun($test_string)
    m = median(t)
    operations_per_second= num*1/(time(m)*1e-9)/1e9
    print("UTF-8: $fun $operations_per_second billion bytes/second\n")
    return (operations_per_second,t)
end
function benchmark_asciivalid(fun; string_length_bytes = 1024)
    
    repeats = div(string_length_bytes, sizeof(ascii_base_string))
    test_string = repeat(ascii_base_string,repeats)
    num = sizeof(test_string)
    t = @benchmark state=$fun($test_string)
    m = median(t)
    operations_per_second= num*1/(time(m)*1e-9)/1e9
    print("ASCII: $fun $operations_per_second billion bytes/second\n")
    return (operations_per_second,t)
end

benchmark_utfvalid(isvalid)
benchmark_asciivalid(isvalid)
```

</details>



<details>
  <summary>Code used to generated dfa table</summary>
This code was taken in large part from  https://github.com/golang/go/issues/47120.

```julia

function build_dfa()
    u0 = zero(UInt64)
	s00 = u0 | (1*6)<<(1*6)
	sC0 = u0
	sC2 = u0 | (2*6)<<(1*6)
	sE0 = u0 | (3*6)<<(1*6)
	sE1 = u0 | (4*6)<<(1*6)
	sED = u0 | (5*6)<<(1*6)
	sEE = sE1
	sF0 = u0 | (6*6)<<(1*6)
	sF1 = u0 | (7*6)<<(1*6)
	sF4 = u0 | (8*6)<<(1*6)
	sF5 = u0

	s80 = u0 | (1*6)<<(2*6) | (2*6)<<(4*6) | (4*6)<<(5*6) | (4*6)<<(7*6) | (4*6)<<(8*6)
	s90 = u0 | (1*6)<<(2*6) | (2*6)<<(4*6) | (4*6)<<(5*6) | (4*6)<<(6*6) | (4*6)<<(7*6)
	sA0 = u0 | (1*6)<<(2*6) | (2*6)<<(3*6) | (2*6)<<(4*6) | (4*6)<<(6*6) | (4*6)<<(7*6)


    dfa = [
        s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00,
        s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00,
        s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00,
        s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00,
        s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00,
        s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00,
        s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00,
        s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00, s00,
        s80, s80, s80, s80, s80, s80, s80, s80, s80, s80, s80, s80, s80, s80, s80, s80,
        s90, s90, s90, s90, s90, s90, s90, s90, s90, s90, s90, s90, s90, s90, s90, s90,
        sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0,
        sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0, sA0,
        sC0, sC0, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2,
        sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2, sC2,
        sE0, sE1, sE1, sE1, sE1, sE1, sE1, sE1, sE1, sE1, sE1, sE1, sE1, sED, sEE, sEE,
        sF0, sF1, sF1, sF1, sF4, sF5, sF5, sF5, sF5, sF5, sF5, sF5, sF5, sF5, sF5, sF5
    ]
    return dfa
end
begin
    old =UInt64(0)
    c = 0
    print("[\n")
    for x in build_dfa()
        if x != old
            c > 0 && print("fill(UInt64($old), $c);\n")
            c = 1
            old = x
        else
            c +=1
        end
    end
    print("fill(UInt64($old), $c)\n")
    print("]\n")
end

```

</details>