@ndinsmore 2022-12-13
Just a further update, it looks like the state table from the reference I had is wrong.  I am working on rebuilding it should be ready tomorrow.