@ndinsmore 2022-12-13
Is the PR the right place to document the build process for the state machine table?