@oscardssmith 2022-12-13
I'd put the table information in a comment in the code.