@ndinsmore 2023-01-06
@stevengj  & @StefanKarpinski (sorry to drag you into this, but I think you did a lot of the initial strings work)
Some of the requested changes(which are 100% correct) beg the question whether this should really be using The `codeunits` interface rather than `unsafe_wrap`.

Initial testing seemed to indicate that `CodeUnits` is slower, but I wonder if it would be wise to do it the correct way here.  Then work on speeding up `CodeUnits` array.