@StefanKarpinski 2023-01-06
Happy to be pulled in! It certainly seems cleaner to use `codeunits`. @JeffBezanson, any idea why that approach would be less efficient?