@ndinsmore 2023-01-11
@stevengj, @StefanKarpinski   & @oscardssmith 
There have been a good amount of changes since the initial PR, but I think it is ready to go. 

I have reversed or order of operations on the shift dfa which leads to state that is never dirty (ie. you would have to ` state & UInt64(63) ` before use) and provided a performance improvement.

Also the use of `codeunits` has dramatically cleaned up the code and made the implementation more maintainable.

With this PR the processing rate is about 0.5 bytes / cycle for short strings (word to line length) and 1.5 bytes/cycle for longer strings ( file length).   This is on par (short) to better (long) than the c implementation.

There are two PRs I plan to follow up with:
1. Broader use of the DFA throughout strings.jl which should simplify the implementation of `length`,`iterate`, `nextind`, and basically anything else that parses through the utf-8 bytes.
2.  A heuristical to approach to short vs long strings.  The current implementation can likely improve short strings 50%, but for long strings, it is fairly easy to get to 10 bytes/cycle for utf-8 and even better 50 bytes/cycle for ascii only strings.

Let me know if there is anything else that needs to be done to this.