@StefanKarpinski 2023-01-27
This is great work. Now that the implementation has moved to Julia, I think we need much more comprehensive tests of the functionality. Previously we could just assume that `byte_string_classify` was correct and simply verify that we use it correctly—basically try one ASCII, one non-ASCII UTF-8 and one invalid string. Now that the implementation is here, we need to have comprehensive tests for it, ideally ones that test all the edge cases of this algorithm, which you're in the best position at this point to understand.