@ndinsmore 2023-02-06
Now that everything goes the the DFA would you agree that if the tests validate the DFA, we can assume `byte_string_classify` is validated?