@ndinsmore 2023-02-06
Test have been added to validate that the DFA state machine returns state as expected and per the Unicode spec.