@ndinsmore 2023-02-10
This can be changed such that the DFA uses a UInt32 table instead of UInt64, similar to what they are doing in rust and postgre.  The downside is that the Table will be made up of magic numbers and much less readable.