@oscardssmith 2023-02-11
if you've seen the `exp` implementation, you would realize that we can be very OK with magic numbers (although you should have a comment with a function that derives them if possible)