@ndinsmore 2023-03-14
This has been updated to use a chunk based system to quickly scan through ascii sections.