@mkitti 2023-03-31
Does DFA stand for Deterministic Finite Automaton?