@ndinsmore 2023-04-06
@mkitti & @oscardssmith  this is a little off topic, but when I rebase I normally just merge master to my local master and rebase to that.  I have had bad luck recently in picking masters that seem to fail tests.  So the question is there a tagged master which is always passing all tests?