@oscardssmith 2023-04-06
not really. in theory, master is supposed to pass tests, but in practice that sometimes doesn't happen.