@ndinsmore 2023-04-11
Is the REPL failure related to this?  I am sure it is text heavy but I don't see any of the changed functions in this PR returning a Union?