@StefanKarpinski 2023-04-12
Exciting times! Was considering how to review this, but it's well tested and I think we'll notice if string stuff is broken.